Refactor alerts to components, test components with fakeAsync

+++ wn-frontend/src/app/admin/admin-two-level-entity/admin-two-level-entity.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit } from '@angular/core'
+import { Component, OnInit, Input } from '@angular/core'
 import { TwoLevelEntityCollection, TwoLevelEntity } from '../../shared/model/two-level-entity.model'
 import { CoreDataService } from '../../core/core-data.service'
 import { ActivatedRoute } from '@angular/router'
+import { NgbModal, ModalDismissReasons, NgbActiveModal } from '@ng-bootstrap/ng-bootstrap'
+import { UtilService } from '../../core/util.service';
 
 @Component({
   selector: 'app-admin-two-level-entity',
@@ -16,8 +18,10 @@ export class AdminTwoLevelEntityComponent implements OnInit {
   sub_item_new: string = ''
 
   constructor(
+    private util: UtilService,
     private coreDataService: CoreDataService,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private modalService: NgbModal
   ) { }
 
   ngOnInit() {
@@ -26,16 +30,14 @@ export class AdminTwoLevelEntityComponent implements OnInit {
   }
 
   loadTwoLevelEntityValues() {
-    return new Promise<TwoLevelEntityCollection>((resolve, reject) => {
-      this.coreDataService.getFlattenedTwoLevelEntityCollection(this.entityName)
-        .then((data) => {
-          this.valuearray = data
-          resolve(data)
-        })
-        .catch((error) => {
-          reject(error)
-        })
-    })
+    this.coreDataService.getFlattenedTwoLevelEntityCollection(this.entityName)
+      .then((data) => {
+        this.valuearray = data
+
+      })
+      .catch((error) => {
+        alert(error.message)
+      })
   }
 
   /**
@@ -43,17 +45,14 @@ export class AdminTwoLevelEntityComponent implements OnInit {
      * to the according array in angular
      */
   createTwoLevelEntityMainItem(newItemName) {
-    return new Promise<TwoLevelEntity>((resolve, reject) => {
-      this.coreDataService.createTwoLevelEntityItem(this.entityName, null, newItemName)
-        .then((data) => {
-          this.loadTwoLevelEntityValues()
-          resolve(data)
-        })
-        .catch((error) => {
-          reject(error)
-        })
-      this.main_item_new = ''
-    })
+    this.coreDataService.createTwoLevelEntityItem(this.entityName, null, newItemName)
+      .then((data) => {
+        this.loadTwoLevelEntityValues()
+      })
+      .catch((error) => {
+        alert(error.message)
+      })
+    this.main_item_new = ''
   }
 
   /**
@@ -61,16 +60,15 @@ export class AdminTwoLevelEntityComponent implements OnInit {
    * on success
    */
   deleteTwoLevelEntityMainItem(item_id) {
-    return new Promise<void>((resolve, reject) => {
+    if (confirm("Eintrag wird gelöscht!")) {
       this.coreDataService.deleteTwoLevelEntityItem(this.entityName, item_id, null)
         .then(() => {
           this.loadTwoLevelEntityValues()
-          resolve()
         })
         .catch((error) => {
-          reject(error)
+          alert(error.message)
         })
-    })
+    }
   }
 
   /**
@@ -80,18 +78,13 @@ export class AdminTwoLevelEntityComponent implements OnInit {
    * // Change to input on first click, then update on second click!
    */
   updateTwoLevelEntityMainItem(item_id, item_edited_name) {
-    return new Promise<TwoLevelEntity>((resolve, reject) => {
-      this.coreDataService.updateTwoLevelEntityItem(this.entityName, item_id, null, item_edited_name)
-        .then((data) => {
-          this.loadTwoLevelEntityValues()
-          resolve(data)
-        })
-        .catch((error) => {
-          reject(error)
-        })
-    })
-
-
+    this.coreDataService.updateTwoLevelEntityItem(this.entityName, item_id, null, item_edited_name)
+      .then((data) => {
+        this.loadTwoLevelEntityValues()
+      })
+      .catch((error) => {
+        alert(error.message)
+      })
   }
 
 
@@ -100,17 +93,14 @@ export class AdminTwoLevelEntityComponent implements OnInit {
      * to the according array in angular
      */
   createTwoLevelEntitySubItem(mainItemId, newItemName) {
-    return new Promise<TwoLevelEntity>((resolve, reject) => {
-      this.coreDataService.createTwoLevelEntityItem(this.entityName, mainItemId, newItemName)
-        .then((data) => {
-          this.loadTwoLevelEntityValues()
-          resolve(data)
-        })
-        .catch((error) => {
-          reject(error)
-        })
-      this.sub_item_new = ''
-    })
+    this.coreDataService.createTwoLevelEntityItem(this.entityName, mainItemId, newItemName)
+      .then((data) => {
+        this.loadTwoLevelEntityValues()
+      })
+      .catch((error) => {
+        alert(error.message)
+      })
+    this.sub_item_new = ''
   }
 
   /**
@@ -118,16 +108,15 @@ export class AdminTwoLevelEntityComponent implements OnInit {
    * on success
    */
   deleteTwoLevelEntitySubItem(mainItemId, subItemId) {
-    return new Promise<any>((resolve, reject) => {
+    if (confirm('Eintrag wird gelöscht!')) {
       this.coreDataService.deleteTwoLevelEntityItem(this.entityName, mainItemId, subItemId)
         .then(() => {
           this.loadTwoLevelEntityValues()
-          resolve()
         })
         .catch((error) => {
-          reject(error)
+          alert(error.message)
         })
-    })
+    }
   }
 
   /**
@@ -137,16 +126,107 @@ export class AdminTwoLevelEntityComponent implements OnInit {
    * // Change to input on first click, then update on second click!
    */
   updateTwoLevelEntitySubItem(mainItemId, subItemId, newItemName) {
-    return new Promise<TwoLevelEntity>((resolve, reject) => {
-      this.coreDataService.updateTwoLevelEntityItem(this.entityName, mainItemId, subItemId, newItemName)
-        .then((data) => {
-          this.loadTwoLevelEntityValues()
-          resolve(data)
+    this.coreDataService.updateTwoLevelEntityItem(this.entityName, mainItemId, subItemId, newItemName)
+      .then((data) => {
+        this.loadTwoLevelEntityValues()
+      })
+      .catch(error => {
+        alert(error.message)
+      })
+  }
+
+  makeMainItem(subItemId) {
+    this.coreDataService.makeMainItem(subItemId, this.entityName)
+      .then((item) => {
+        this.util.removeFromArray(this.valuearray.values, item)
+        this.loadTwoLevelEntityValues()
+      })
+      .catch((error) => {
+        alert(error.message)
+      })
+  }
+
+  makeSubItem(item) {
+    if (item && item.id) {
+      // Open Modal to select new main item
+      console.log('addAsSubItem() called')
+      this.openAddAsSubItemModal(item)
+        .then((mainItem) => {
+          // Send request
+          this.addAsSubItemRequest(item.id, mainItem.id, this.entityName)
+            .then((data) => {
+              this.util.removeFromArray(this.valuearray.values, item)
+              this.loadTwoLevelEntityValues()
+            })
+            .catch((error) => {
+              alert(error.message)
+            })
         })
-        .catch(error => {
+        .catch((error) => {
+          // Do nothing, window is only dismissed
+          console.log(error)
+        })
+    } else {
+      alert('Es wurde kein Item mit Id übergeben!')
+    }
+  }
+
+  openAddAsSubItemModal(item) {
+    return new Promise<TwoLevelEntity>((resolve, reject) => {
+      let modalRef = this.modalService.open(MakeSubItemModalContent)
+      modalRef.componentInstance.valuearray = this.valuearray
+      modalRef.componentInstance.item = item
+
+      modalRef.result.then(
+        (result) => {
+          resolve(result)
+        },
+        (reason) => {
+          console.log('DISMISSED')
+          reject(reason)
+        });
+    })
+  }
+
+  addAsSubItemRequest(itemId, mainItemId, entityName) {
+    return new Promise<TwoLevelEntity>((resolve, reject) => {
+      this.coreDataService.addAsSubItem(itemId, mainItemId, entityName)
+        .then((item) => {
+          resolve(item)
+        })
+        .catch((error) => {
           reject(error)
         })
     })
   }
+}
 
+@Component({
+  selector: 'make-sub-item-modal-content',
+  template: `
+    <div class="modal-header">
+      <h4 class="modal-title">Neues Hauptitem für {{ item.name }} wählen</h4>
+    </div>
+    <div class="modal-body">
+      <select class="form-control" name="select-new-main-item" [(ngModel)]="newMainItem" [compareWith]='util.compareById'>
+        <option [value]='undefined' selected>-- Bitte wählen --</option>
+        <option *ngFor="let option of valuearray.values" [ngValue]="option">{{ option.name }}</option>
+      </select>
+    </div>
+    <div class="modal-footer">
+      <button type="button" class="btn btn-default" (click)="activeModal.dismiss()">Abbrechen</button>
+      <button type="button" class="btn btn-primary" (click)="activeModal.close(newMainItem)">Speichern</button>
+    </div>
+  `
+})
+export class MakeSubItemModalContent {
+  @Input() item
+  @Input() valuearray
+
+  public newMainItem: TwoLevelEntity
+
+  constructor(
+    public util: UtilService,
+    public activeModal: NgbActiveModal
+  ) { }
 }
